requirements.txt
apipkg==1.5
asttokens==2.0.3
attrdict==2.0.1
attrs==19.3.0
base58==2.0.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
cycler==0.10.0
cytoolz==0.10.1
decorator==4.4.2
eth-abi==2.1.1
eth-account==0.4.0
eth-brownie==1.7.5
eth-event==1.0.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.4
execnet==1.7.1
future==0.18.2
hexbytes==0.2.0
hypothesis==5.10.1
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
ipfshttpclient==0.4.13.2
jsonschema==3.2.0
kiwisolver==1.2.0
lru-dict==1.1.6
matplotlib==3.2.1
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==8.3.0
multiaddr==0.0.9
mythx-models==1.7.7
netaddr==0.7.19
numpy==1.18.1
packaging==20.4
pandas==1.0.3
parsimonious==0.8.1
pluggy==0.13.1
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
py==1.8.1
py-solc-ast==1.2.4
py-solc-x==0.8.2
pycparser==2.20
pycryptodome==3.9.7
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.1
pytest-forked==1.1.3
pytest-xdist==1.31.0
python-dateutil==2.8.1
python-telegram-bot==12.7
pythx==1.5.6
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
rlp==1.2.0
semantic-version==2.8.5
six==1.15.0
sortedcontainers==2.1.0
toolz==0.10.0
tornado==6.0.4
tqdm==4.45.0
typing-extensions==3.7.4.2
urllib3==1.25.9
varint==1.0.2
vyper==0.1.0b17
wcwidth==0.1.9
web3==5.5.1
websockets==8.1
zipp==3.1.0
